this update reads in the contents of the config files rather than opening them as a stream. we've also wrapped the saves in a try/catch (not 100% sure if this'll cause usync to fail silently or will display the error in the usync developer tab?).
the main reason for this is we have environments we upload too that are locked down and we cannot open files using system.io due to permissions. however, in that example we upload existing config files that usync can read in as we don't want the production servers writing to the the configs, we only want them to read.