Container Folder update
Didn't need to use Guid Keys for container items, this update lets the container folders be created on demand. (this might cause an issue for deletes i need to confirm). So we no longer create an export just for the container .

This works, but need to think about it, because it might be better to create the containers and let usync handle them as such, just because it will also handle deletes then. using the above we are going to have to write some speical delete code to capture the empty container folders (we could have a post processing handler clean up empty ones - but what if you want empty ones?)