Fix Archetype same-alias property issue (#359)
When there are multiple Archetype fieldset types which contain the same property alias, this causes the property value to get mapped multiple times. For an media picker property, this causes the guids to get distorted.

This change ensures that the mapping is only applied a single time for each fieldset and property.